HIVE-24495: Extract Thrift server launch logic to ThriftMetastoreLauncher
