+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/HMSHandler.java
@@ -1028,7 +1028,7 @@ public class HMSHandler extends FacebookBase implements IHMSHandler {
 
   @Override
   public void shutdown() {
-    HiveMetaStore.cleanupRawStore();
+    ThriftMetastoreLauncher.cleanupRawStore();
     PerfLogger.getPerfLogger(false).cleanupPerfLogMetrics();
     ThreadPool.shutdown();
   }
