+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/MetastoreLauncher.java
@@ -1,0 +1,7 @@
+package org.apache.hadoop.hive.metastore;
+
+import org.apache.hadoop.conf.Configuration;
+
+public interface MetastoreLauncher extends Runnable{
+  MetastoreLauncher configure(Configuration conf, HiveMetaStore.HiveMetastoreCli cli);
+}
